# speedup shift opertions