Using the C level shift operator, this PR speeds up Python bitarray shift operations (`<<`, `>>`, `<<=`, `>>=`) by a large amount.  This is done by shifting 64 bit words when possible and chars directly.  Here are measurements for shitting a bitarray of size 2^30, by one bit inplace:
```
 0.092 sec   little-endian  left-shift
 0.029 sec   little-endian  right-shift
 0.263 sec   big-endian  left-shift
 0.168 sec   big-endian  right-shift
```
Without this PR:
```
 2.537 sec   little-endian  left-shift
 2.467 sec   little-endian  right-shift
 2.515 sec   big-endian  left-shift
 2.471 sec   big-endian  right-shift
```