@ilanschnell 2021-07-20
This is slightly related to #137, as I which to gain more experience with shift operations first.